feat: refactor session management to use NewComputeSession component and add new data volume and user volume accordions
