+++ components/ui/web/components/content/newComputeSession/workingDirectoryAccordion.tsx
@@ -2,7 +2,7 @@ import { FC, useState } from 'react';
 import { AccordionDetails } from '@mui/material';
 import styled from 'styled-components';
 
-import { ParamAccordionSummary } from 'components/content/newResource/paramAccordionSummary';
+import { ParamAccordionSummary } from 'components/content/newComputeSession/paramAccordionSummary';
 import { UserVolume } from 'src/graphql/typings';
 import { StyledAccordion } from 'components/common/accordion';
 import { InfoCard } from 'components/common/infoCard';
